refactor: refactor book, sectons, styles
