+++ src/app/const.ts
@@ -1,1 +1,2 @@
 export const SITE_ORIGIN = 'http://localhost:5000/';
+export const WORD_PER_PAGE = 19;
