+++ src/styles/pages/_book.scss
@@ -1,13 +1,6 @@
-$font-family: "Neucha", cursive;
-$font-color: #455d7a;
-
-body {
-  background-color: $gray_light;
-}
-
 .book-header {
-  font-family: $font-family;
-  color: $font-color;
+  font-family: $font-ui;
+  color: $navy;
   font-weight: 600;
   font-size: 30px;
   padding: 10px;
@@ -39,11 +32,10 @@ body {
       @include shadow;
       a {
         text-decoration: none;
-        font-family: $font-family;
-        color: $font-color;
+        font-family: $font-ui;
+        color: $black;
         font-size: 24px;
         line-height: 28px;
-        letter-spacing: 2px;
       }
     }
   }
@@ -53,7 +45,7 @@ body {
   width: 55%;
   margin-top: 5px;
   font-family: $font-c;
-  color: $font-color;
+  color: $black;
   font-size: $font_m;
   letter-spacing: 1px;
   background-color: $white;
@@ -78,11 +70,11 @@ body {
   font-size: 20px;
 }
 
-.fa-check-circle {
+.learned {
   color: $green_dark;
 }
 
-.fa-star {
+.difficult {
   color: $red;
 }
